renamed files, cleaning up code
